Add pagination and transclusion
Auto-indexed directories now generate pagination if the following query
parameters are present in the URL:

- p: page index starting at 1
- n: number of entries per page

HTML files now support simple transclusion: if the pattern {FILE-PATH}
is encountered, the contents of the file at FILE-PATH will be included
in-place wherever the pattern occured. { and } can be escaped as \{ \}
to produce a literal { } without transclusion.

Other changes:
 - moved micro-http-server into the project dir so it can be easily
   modified.
 - automatically redirect to append a "/" to directory URLs